requirements.txt
pandas
pytorch torchvision torchaudio cudatoolkit=11.0 -c pytorch
pytorch_lightning
transformers
soynlp emoji